+++ routes/routes.go
@@ -2,6 +2,7 @@ package routes
 
 import (
 	"github.com/Bluhabit/uwang-rest-account/middlewares"
+	"github.com/Bluhabit/uwang-rest-account/routes/user"
 	"github.com/Bluhabit/uwang-rest-account/routes/profile"
 	"github.com/gin-gonic/gin"
 )
@@ -13,6 +14,6 @@ func InitRoutes(router *gin.Engine) {
 		v1.POST("/update-profile-picture", middlewares.AuthMiddleware(), profile.UpdateProfilePicture)
 		v1.POST("/update-profile-interest-topics", middlewares.AuthMiddleware(), profile.UpdateProfileInterestTopics)
 		v1.POST("/update-profile-level", middlewares.AuthMiddleware(), profile.UpdateProfileLevel)
-		v1.GET("/users",profile.DetailUserResponse)
+		v1.GET("/users",user.DetailUserResponse)
 	}
 }
